requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.3
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.10.0=pypi_0
aom=3.9.1=h7bae524_0
appnope=0.1.4=pypi_0
asttokens=3.0.0=pypi_0
attrs=25.3.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
bleach=6.2.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.7.14=hbd8a1cb_0
cachetools=5.5.2=pypi_0
cairo=1.18.4=h6a3b0d2_0
certifi=2025.7.14=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.1.8=pypi_0
dav1d=1.2.1=hb547adb_0
dbus=1.13.6=h3818c69_3
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
distro=1.9.0=pypi_0
docopt=0.6.2=pypi_0
executing=2.2.1=pypi_0
expat=2.7.1=h313beb8_0
fastjsonschema=2.21.2=pypi_0
ffmpeg=7.1.1=gpl_hf33b5e2_107
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h1383a14_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.13.3=hce30654_1
fribidi=1.0.10=h27ca646_0
gdk-pixbuf=2.42.12=h7ddc832_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
gmp=6.3.0=h7bae524_2
google-ai-generativelanguage=0.6.15=pypi_0
google-api-core=2.25.1=pypi_0
google-api-python-client=2.177.0=pypi_0
google-auth=2.40.3=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.2.2=pypi_0
google-cloud-texttospeech=2.27.0=pypi_0
google-generativeai=0.8.5=pypi_0
googleapis-common-protos=1.70.0=pypi_0
graphite2=1.3.14=h286801f_0
grpcio=1.74.0=pypi_0
grpcio-status=1.71.2=pypi_0
grpcio-tools=1.63.2=pypi_0
gtts=2.5.4=pypi_0
h11=0.16.0=pypi_0
harfbuzz=11.1.0=hab40de2_0
httpcore=1.0.9=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.28.1=pypi_0
icu=75.1=hfee45f7_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
imageio-ffmpeg=0.6.0=pypi_0
instagrapi=2.2.1=pypi_0
ipython=8.12.3=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.10.0=pypi_0
jsonschema=4.25.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.8.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
lame=3.100=h1a8c8d9_1003
lerc=4.0.0=hd64df32_1
libabseil=20250512.1=cxx17_hd41c47c_0
libass=0.17.4=hcbd7ca7_0
libcxx=20.1.8=hf598326_1
libdeflate=1.24=h5773f1b_0
libexpat=2.7.1=hec049ff_0
libffi=3.4.4=hca03da5_1
libfreetype=2.13.3=hce30654_1
libfreetype6=2.13.3=h1d14073_1
libglib=2.84.0=hdff4504_0
libhwloc=2.12.1=default_h88f92a7_1000
libiconv=1.18=hfe07756_1
libintl=0.25.1=h493aca8_0
libjpeg-turbo=3.1.0=h5505292_0
liblzma=5.8.1=h39f12f2_2
liblzma-devel=5.8.1=h39f12f2_2
libogg=1.3.5=h48c0fde_1
libopenvino=2025.2.0=h56e7ac4_1
libopenvino-arm-cpu-plugin=2025.2.0=h56e7ac4_1
libopenvino-auto-batch-plugin=2025.2.0=he81eb65_1
libopenvino-auto-plugin=2025.2.0=he81eb65_1
libopenvino-hetero-plugin=2025.2.0=h273c05f_1
libopenvino-ir-frontend=2025.2.0=h273c05f_1
libopenvino-onnx-frontend=2025.2.0=h6386500_1
libopenvino-paddle-frontend=2025.2.0=h6386500_1
libopenvino-pytorch-frontend=2025.2.0=hec049ff_1
libopenvino-tensorflow-frontend=2025.2.0=hee62d61_1
libopenvino-tensorflow-lite-frontend=2025.2.0=hec049ff_1
libopus=1.5.2=h48c0fde_0
libpng=1.6.50=h3783ad8_0
libprotobuf=6.31.1=h702a38d_1
librsvg=2.58.4=h266df6f_3
libsqlite=3.45.2=h091b4b1_0
libtiff=4.7.0=h2f21f7c_5
libusb=1.0.29=hbc156a2_0
libvorbis=1.3.7=h9f76cd9_0
libvpx=1.14.1=h7bae524_0
libwebp-base=1.6.0=h07db88b_0
libxml2=2.13.8=h52572c6_0
libzlib=1.3.1=h8359307_2
markupsafe=3.0.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.1.4=pypi_0
moviepy=1.0.3=pypi_0
narwhals=2.0.1=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h313beb8_0
numpy=2.3.2=pypi_0
oauthlib=3.3.1=pypi_0
openai=1.99.9=pypi_0
openh264=2.6.0=hb5b2745_0
openssl=3.5.1=h81ee809_0
packaging=25.0=pypi_0
pandas=2.3.1=pypi_0
pandocfilters=1.5.1=pypi_0
pango=1.56.3=h5fd7515_1
param=2.2.1=pypi_0
parso=0.8.5=pypi_0
pcre2=10.44=ha881caa_2
pexels-api=1.0.1=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=25.1=pyhc872135_2
pipreqs=0.5.0=pypi_0
pixman=0.46.4=h2c80e29_0
platformdirs=4.4.0=pypi_0
proglog=0.1.12=pypi_0
prompt-toolkit=3.0.52=pypi_0
proto-plus=1.26.1=pypi_0
protobuf=5.29.5=pypi_0
ptyprocess=0.7.0=pypi_0
pugixml=1.15=hd3d436d_0
pure-eval=0.2.3=pypi_0
pyarrow=21.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pycryptodomex=3.23.0=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.19.2=pypi_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=pypi_0
python=3.11.8=hdf0ec26_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.1=pypi_0
pytz=2025.2=pypi_0
pyzmq=27.0.2=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.36.2=pypi_0
requests=2.32.4=pypi_0
requests-oauthlib=2.0.0=pypi_0
rpds-py=0.27.0=pypi_0
rsa=4.9.1=pypi_0
schedule=1.2.2=pypi_0
sdl2=2.32.54=ha1acc90_0
sdl3=3.2.14=hf196eef_0
setuptools=78.1.1=py311hca03da5_0
six=1.17.0=pypi_0
smmap=5.0.2=pypi_0
snappy=1.2.2=hd121638_0
sniffio=1.3.1=pypi_0
soupsieve=2.7=pypi_0
sqlite=3.45.2=hf2abe2d_0
stack-data=0.6.3=pypi_0
streamlit=1.48.0=pypi_0
svt-av1=3.0.2=h8ab69cd_0
tbb=2022.2.0=h5b2e6d4_0
tenacity=9.1.2=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.13=h892fb3f_2
toml=0.10.2=pypi_0
tornado=6.5.1=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
typing-extensions=4.14.1=pypi_0
typing-inspection=0.4.1=pypi_0
tzdata=2025.2=pypi_0
uritemplate=4.2.0=pypi_0
urllib3=2.5.0=pypi_0
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.45.1=py311hca03da5_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xz=5.8.1=h9a6d368_2
xz-gpl-tools=5.8.1=h9a6d368_2
xz-tools=5.8.1=h39f12f2_2
yarg=0.1.9=pypi_0
zlib=1.3.1=h8359307_2
zstd=1.5.7=h6491c7d_2
